Improve logic to register and unregister hub topic while saving and deleting websub configuration
