+++ components/notification-sender-config/org.wso2.carbon.identity.notification.sender.tenant.config/src/main/java/org/wso2/carbon/identity/notification/sender/tenant/config/handlers/DefaultChannelChannelConfigurationHandler.java
@@ -33,14 +33,18 @@ import org.wso2.carbon.identity.configuration.mgt.core.exception.ConfigurationMa
 import org.wso2.carbon.identity.configuration.mgt.core.model.Attribute;
 import org.wso2.carbon.identity.configuration.mgt.core.model.Resource;
 import org.wso2.carbon.identity.configuration.mgt.core.model.ResourceFile;
+import org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants;
 import org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementServiceImpl;
+import org.wso2.carbon.identity.notification.sender.tenant.config.clustering.EventPublisherClusterDeleteMessage;
 import org.wso2.carbon.identity.notification.sender.tenant.config.clustering.EventPublisherClusterInvalidationMessage;
 import org.wso2.carbon.identity.notification.sender.tenant.config.dto.SMSSenderDTO;
 import org.wso2.carbon.identity.notification.sender.tenant.config.exception.NotificationSenderManagementClientException;
 import org.wso2.carbon.identity.notification.sender.tenant.config.exception.NotificationSenderManagementException;
 import org.wso2.carbon.identity.notification.sender.tenant.config.exception.NotificationSenderManagementServerException;
 import org.wso2.carbon.identity.notification.sender.tenant.config.internal.NotificationSenderTenantConfigDataHolder;
+import org.wso2.carbon.identity.tenant.resource.manager.exception.TenantResourceManagementClientException;
 import org.wso2.carbon.identity.tenant.resource.manager.exception.TenantResourceManagementException;
+import org.wso2.carbon.identity.tenant.resource.manager.exception.TenantResourceManagementServerException;
 import org.wso2.carbon.identity.tenant.resource.manager.util.ResourceUtils;
 
 import java.io.InputStream;
@@ -58,6 +62,7 @@ import static org.wso2.carbon.identity.notification.sender.tenant.config.Notific
 import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.DEFAULT_SMS_PUBLISHER;
 import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.ErrorMessage.ERROR_CODE_CONFLICT_PUBLISHER;
 import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.ErrorMessage.ERROR_CODE_ERROR_ADDING_NOTIFICATION_SENDER;
+import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.ErrorMessage.ERROR_CODE_ERROR_DELETING_NOTIFICATION_SENDER;
 import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.ErrorMessage.ERROR_CODE_NO_ACTIVE_PUBLISHERS_FOUND;
 import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.ErrorMessage.ERROR_CODE_PARSER_CONFIG_EXCEPTION;
 import static org.wso2.carbon.identity.notification.sender.tenant.config.NotificationSenderManagementConstants.ErrorMessage.ERROR_CODE_PUBLISHER_NOT_EXISTS_IN_SUPER_TENANT;
@@ -130,6 +135,23 @@ public class DefaultChannelChannelConfigurationHandler extends ChannelConfigurat
             throw handleConfigurationMgtException(e, ERROR_CODE_ERROR_ADDING_NOTIFICATION_SENDER, smsSender.getName());
         }
         return buildSmsSenderFromResource(smsSenderResource);
+    }
+
+    @Override
+    public void deleteNotificationSender(String senderName) throws NotificationSenderManagementException {
+
+        try {
+            NotificationSenderTenantConfigDataHolder.getInstance().getResourceManager()
+                    .removeEventPublisherConfiguration(PUBLISHER_RESOURCE_TYPE, senderName);
+            NotificationSenderTenantConfigDataHolder.getInstance().getConfigurationManager()
+                    .deleteResource(PUBLISHER_RESOURCE_TYPE, senderName);
+            sendEventPublisherClusterDeleteMessage(senderName);
+
+        } catch (ConfigurationManagementException e) {
+            throw handleConfigurationMgtException(e, ERROR_CODE_ERROR_DELETING_NOTIFICATION_SENDER, senderName);
+        } catch (TenantResourceManagementException e) {
+            throw handleTenantResourceManagementException(e, ERROR_CODE_ERROR_DELETING_NOTIFICATION_SENDER, senderName);
+        }
     }
 
     private void validateSMSSender(SMSSenderDTO smsSender) throws NotificationSenderManagementClientException {
@@ -312,6 +334,33 @@ public class DefaultChannelChannelConfigurationHandler extends ChannelConfigurat
                     // Do nothing.
                 }
             }
+        }
+    }
+
+    private void sendEventPublisherClusterDeleteMessage(String senderName) {
+
+        if (getClusteringAgent() == null) {
+            return;
+        }
+
+        EventPublisherClusterDeleteMessage message = new EventPublisherClusterDeleteMessage(
+                NotificationSenderManagementConstants.PUBLISHER_RESOURCE_TYPE, senderName,
+                PrivilegedCarbonContext.getThreadLocalCarbonContext().getTenantId());
+
+        sendClusterMessage(message, senderName);
+    }
+
+    private NotificationSenderManagementException handleTenantResourceManagementException(
+            TenantResourceManagementException e,
+            NotificationSenderManagementConstants.ErrorMessage error, String data) {
+
+
+        if (e instanceof TenantResourceManagementClientException) {
+            return new NotificationSenderManagementClientException(error, data, e);
+        } else if (e instanceof TenantResourceManagementServerException) {
+            return new NotificationSenderManagementServerException(error, data, e);
+        } else {
+            return new NotificationSenderManagementException(error, data, e);
         }
     }
 }
